Sync OpenBSD patchset 991:
Extend history that can be captured to INT_MIN rather than
SHRT_MIN. From Chris Johnsen.

+++ cmd-capture-pane.c
@@ -59,7 +59,7 @@ cmd_capture_pane_exec(struct cmd *self, struct cmd_ctx *ctx)
 	buf = NULL;
 	len = 0;
 
-	n = args_strtonum(args, 'S', SHRT_MIN, SHRT_MAX, &cause);
+	n = args_strtonum(args, 'S', INT_MIN, SHRT_MAX, &cause);
 	if (cause != NULL) {
 		top = gd->hsize;
 		xfree(cause);
@@ -70,7 +70,7 @@ cmd_capture_pane_exec(struct cmd *self, struct cmd_ctx *ctx)
 	if (top > gd->hsize + gd->sy - 1)
 		top = gd->hsize + gd->sy - 1;
 
-	n = args_strtonum(args, 'E', SHRT_MIN, SHRT_MAX, &cause);
+	n = args_strtonum(args, 'E', INT_MIN, SHRT_MAX, &cause);
 	if (cause != NULL) {
 		bottom = gd->hsize + gd->sy - 1;
 		xfree(cause);
